# docs/swagger.yaml
swagger: '2.0'
info:
  title: Predico - Restful-API
  description: |

    #### Description:
    Restful API for the INESC TEC Data Sharing / Barter Incentives Service.

    This API is crafted to enable secure data exchange and incentivized collaboration among multiple agents and a central server, where the execution of Data Sharing/Barter Incentives algorithms takes place. It robustly supports internal mechanisms for Data Sharing/Barter Incentives, addressing a diverse range of use cases.

    A prominent use case, characterized by a higher Technology Readiness Level (TRL), is the collaborative forecasting scenario.
    In this context, various agents can trade historical (or future) time-series data—such as power generation or weather measurements data—and acquire time-series forecasts improved by the collective pool of data available on the platform. Notably, agents can access these forecasts without directly tapping into internal datasets. The economic value of data shared by one agent is inherently tied to its utility for other participants that also buy collaborative forecasts from this service.

    ## Main functions:
    - **User management**:
        * Register new user
        * User verification / authentication
        * User password reset
        * Manage User Resources (i.e., bidding units in user portfolio) (create/update/delete)
        * Manage User IOTA wallet address reference (create/update/delete)

    - **Market management**:
...

...

...
  license:
...
  version: v1
host: predico02.inesctec.pt
schemes:
...
basePath: /api
consumes:
...
produces:
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
paths:
...
  /user/resource/:
    get:
      operationId: get_user_resources
      description: Method for agents to list their registered resources
      parameters:
        - name: resource
          in: query
          description: Filter by resource identifier
          required: false
          type: integer
        - name: resource_name
          in: query
          description: Filter by resource name
          required: false
          type: string
        - name: to_forecast
          in: query
          description: Return resources currently registered as 'to_forecast' in the market
          required: false
          type: boolean
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 200
              data:
                description: Response data.
                type: object
          examples:
            application/json:
              code: 200
              data:
                - id: 1
                  name: park-1
                  type: measurements
                  to_forecast: true
                  registered_at: '2022-10-20T11:34:57.739206Z'
                  user: 1
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - user
    post:
      operationId: post_user_resources
      description: Method for agents to register a new resource
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserResources'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/UserResources'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - user
    parameters: []
  /user/resource/{resource_id}:
    patch:
      operationId: patch_user_resources
      description: Method for agents to update their resource details
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserResources'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/UserResources'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - user
    delete:
      operationId: user_resource_delete
      description: ''
      parameters: []
      responses:
        '204':
          description: ''
      tags:
        - user
    parameters:
      - name: resource_id
        in: path
        required: true
        type: string
...
definitions:
...
  UserResources:
    required:
      - name
      - type
      - to_forecast
      - user
    type: object
    properties:
      id:
        title: Id
        type: integer
        readOnly: true
      name:
        title: Name
        type: string
        pattern: ^[a-z0-9_-]*$
        maxLength: 64
        minLength: 1
      type:
        title: Type
        type: string
        enum:
          - forecasts
          - measurements
      to_forecast:
        title: To forecast
        type: boolean
      registered_at:
        title: Registered at
        type: string
        format: date-time
        readOnly: true
      user:
        title: User
        type: integer
...